refactor: moved cli and daemon into their own modules
